pyproject.toml
[tool.poetry]
name = "frost"
version = "0.1.1"
description = ""
authors = ["Ole Martin Ruud <dev@barskern.no>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pendulum = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
autopep8 = "^2.3.1"
flake8 = "^7.1.1"
autoflake8 = "^0.4.0"
pre-commit = "^4.0.1"
requests-cache = "^1.2.1"

[tool.poetry.scripts]
frost_fetch = "frost:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
